requirements.txt
numpy
cython
gensim
pyvi
scikit-surprise
dill
pandas
streamlit